krux-shorts/requirements.txt
anthropic
openai
supabase
fal-client
python-dotenv
requests
